refactored customisations module to be multiplatform friendly
NOTE: JS implementation is now a no-op situation
