+++ utils/customisations/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     kotlin("multiplatform")
+    id("com.android.library")
     id("appyx-publish-multiplatform")
     id("appyx-detekt")
 }
@@ -9,7 +10,10 @@ publishingPlugin {
 }
 
 kotlin {
-    jvm {
+    android {
+        publishLibraryVariants("release")
+    }
+    jvm("desktop") {
         compilations.all {
             kotlinOptions.jvmTarget = libs.versions.jvmTarget.get()
         }
@@ -21,5 +25,18 @@ kotlin {
     }
     sourceSets {
         val commonMain by getting
+        val androidMain by getting
+        val desktopMain by getting
+        val jsMain by getting
+    }
+}
+
+android {
+    namespace = "com.bumble.appyx.utils.multiplatform"
+    compileSdk = libs.versions.androidCompileSdk.get().toInt()
+    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
+    defaultConfig {
+        minSdk = libs.versions.androidMinSdk.get().toInt()
+        targetSdk = libs.versions.androidTargetSdk.get().toInt()
     }
 }
