+++ utils/customisations/src/jsMain/kotlin/com/bumble/appyx/utils/customisations/NodeCustomisationDirectoryImpl.kt
@@ -1,0 +1,33 @@
+package com.bumble.appyx.utils.customisations
+
+import kotlin.reflect.KClass
+
+actual open class NodeCustomisationDirectoryImpl(
+    override val parent: NodeCustomisationDirectory? = null
+) : MutableNodeCustomisationDirectory {
+
+    override fun <T : NodeCustomisation> put(key: KClass<T>, value: T) {
+        // NO-OP
+    }
+
+    override fun <T : NodeCustomisation> get(key: KClass<T>): T? =
+        null
+
+    override fun <T : NodeCustomisation> getRecursively(key: KClass<T>): T? =
+        null
+
+    override fun <T : Any> putSubDirectory(key: KClass<T>, value: NodeCustomisationDirectory) {
+        // NO-OP
+    }
+
+    override fun <T : Any> getSubDirectory(key: KClass<T>): NodeCustomisationDirectory? =
+        null
+
+    override fun <T : Any> getSubDirectoryOrSelf(key: KClass<T>): NodeCustomisationDirectory {
+        return this
+    }
+
+    operator fun KClass<*>.invoke(block: NodeCustomisationDirectoryImpl.() -> Unit) {
+        // NO-OP
+    }
+}
